feat(dropdown,search): highlightMatches option
This PR adds a new option highlightMatches to dropdown and search module
Customizing the mark (background)-color is prepared in the LESS files for a custom theme, but disabled by default. (as the default would do nothing anyway)
I decided to leave this to each browsers defaults as its visuals are most familiar to usual browser search and the mark HTML tag uses the exact same styling.

Also fixed some typo inside search.less
Also made ignoreSearchCase available for search as we have the same for dropdown already