feat(tab): no "x-remote" header for ajax requests by default
This is documented in the docs - fomantic/Fomantic-UI-Docs@5ad17488bc/server/documents/modules/tab.html.eco#L644.

But I believe this concept is not unified across the whole Fomantic UI and forcing an extra HTTP header is not wanted in general.

Other components, like dropdown and search, do not use this header as the related backend for such components is supposed to always being used for ajax requests, but tab content could be whole HTML code, so the extra header was implemented only for tabs